logical changes

+++ Practice/equilibriumPoint.cpp
@@ -33,7 +33,7 @@ void solve()
     int n;
     cin >> n;
 
-    int arr[n];
+    long long arr[n];
     for (int i = 0; i < n; i++)
         cin >> arr[i];
